Extract markers into pytest_pyvista/_markers.py
Moves the skip_egl/skip_mac/skip_windows/needs_vtk_version markers and
their tests out of pytest_pyvista.py/test_pyvista.py into a dedicated
_markers.py/test_markers.py pair, mirroring the doc_mode.py/test_doc_mode.py
split already used in this repo. Pure move: pytest_pyvista.py now imports
register_markers and re-exports pytest_runtest_setup so pytest still
discovers the hook on the registered plugin module.
